Move popup pill/icon buttons into the shared design-system components
PopupButton and IconButton were popup-local but already reused by
ManageAllStyles and ReleaseNotification, so promote them into
@stylebot/components alongside the other shared primitives.
