+++ src/components/index.ts
@@ -2,3 +2,5 @@ export { default as Heading } from './Heading.vue';
 export { default as TextBlock } from './TextBlock.vue';
 export { default as ShortcutChip } from './ShortcutChip.vue';
 export { default as ShortcutKbd } from './ShortcutKbd.vue';
+export { default as PopupButton } from './PopupButton.vue';
+export { default as IconButton } from './IconButton.vue';
